# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/books": {
      "get": {
        "description": "Retrieves a paginated list of books with support for search, filtering, and sorting. Public endpoint accessible without authentication.",
        "operationId": "BooksController_findAll",
        "parameters": [
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order (default: asc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field (default: relevance if searching, createdAt otherwise)",
            "schema": {
              "enum": [
                "title",
                "createdAt",
                "relevance"
              ],
              "type": "string"
            }
          },
          {
            "name": "availability",
            "required": false,
            "in": "query",
            "description": "Filter by availability status",
            "schema": {
              "enum": [
                "AVAILABLE",
                "BORROWED",
                "RESERVED",
                "ALL"
              ],
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter by category UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "description": "Filter by author UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query for title and author name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10, max: 100)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Books retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "books": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": "string",
                            "example": "The Great Gatsby"
                          },
                          "subtitle": {
                            "type": "string",
                            "nullable": true
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "isbn": {
                            "type": "string",
                            "example": "978-0743273565"
                          },
                          "publicationYear": {
                            "type": "number",
                            "nullable": true,
                            "example": 1925
                          },
                          "language": {
                            "type": "string",
                            "nullable": true,
                            "example": "English"
                          },
                          "coverImageUrl": {
                            "type": "string",
                            "format": "uri",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "ACTIVE",
                              "ARCHIVED"
                            ],
                            "example": "ACTIVE"
                          },
                          "authors": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          },
                          "categories": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          },
                          "availableCopies": {
                            "type": "number",
                            "example": 3
                          },
                          "totalCopies": {
                            "type": "number",
                            "example": 5
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "pageSize": {
                          "type": "number",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 100
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 10
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          }
        },
        "summary": "List all books",
        "tags": [
          "Books"
        ]
      },
      "post": {
        "description": "Creates a new book in the catalog. Admin only. Validates ISBN format and uniqueness.",
        "operationId": "BooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Book data",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "isbn",
                  "authorIds",
                  "categoryIds"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "The Great Gatsby"
                  },
                  "subtitle": {
                    "type": "string",
                    "maxLength": 500,
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "isbn": {
                    "type": "string",
                    "example": "978-0743273565",
                    "description": "ISBN-10 or ISBN-13"
                  },
                  "publicationYear": {
                    "type": "number",
                    "minimum": 1000,
                    "nullable": true,
                    "example": 1925
                  },
                  "language": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true,
                    "example": "English"
                  },
                  "coverImageUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "authorIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully"
          },
          "400": {
            "description": "Invalid input data or ISBN already exists"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (ADMIN only)"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new book",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/books/{id}": {
      "get": {
        "description": "Retrieves detailed information about a specific book including authors, categories, and copy availability.",
        "operationId": "BooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Book UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "book": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "title": {
                          "type": "string"
                        },
                        "subtitle": {
                          "type": "string",
                          "nullable": true
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "isbn": {
                          "type": "string"
                        },
                        "publicationYear": {
                          "type": "number",
                          "nullable": true
                        },
                        "language": {
                          "type": "string",
                          "nullable": true
                        },
                        "coverImageUrl": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "ARCHIVED"
                          ]
                        },
                        "authors": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "categories": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "availableCopies": {
                          "type": "number"
                        },
                        "totalCopies": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found or archived"
          }
        },
        "summary": "Get a book by ID",
        "tags": [
          "Books"
        ]
      },
      "patch": {
        "description": "Updates an existing book. Admin only. All fields are optional.",
        "operationId": "BooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Book UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Book data to update (all fields optional)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "subtitle": {
                    "type": "string",
                    "maxLength": 500,
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "isbn": {
                    "type": "string",
                    "description": "ISBN-10 or ISBN-13"
                  },
                  "publicationYear": {
                    "type": "number",
                    "minimum": 1000,
                    "nullable": true
                  },
                  "language": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true
                  },
                  "coverImageUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "ARCHIVED"
                    ]
                  },
                  "authorIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (ADMIN only)"
          },
          "404": {
            "description": "Book not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a book",
        "tags": [
          "Books"
        ]
      },
      "delete": {
        "description": "Permanently deletes a book if it has no loan history. If loans exist, the book should be archived instead. Admin only.",
        "operationId": "BooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Book UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (ADMIN only)"
          },
          "404": {
            "description": "Book not found"
          },
          "409": {
            "description": "Book has loan history and cannot be deleted. Archive it instead."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a book",
        "tags": [
          "Books"
        ]
      }
    },
...
  },
  "info": {
    "title": "Library Management System API",
    "description": "RESTful API for a comprehensive Library Management System. This API provides endpoints for managing books, authors, categories, book copies, members, loans, and system settings.",
    "version": "1.0.0",
    "contact": {
      "name": "Library Management Team",
      "url": "https://github.com/library-management",
      "email": "support@library-management.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.library-management.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}